Fractional Qty is realised via aliases

+++ src/toy_trader/strategies.py
@@ -67,14 +67,15 @@ class SMACrossStrategy(Strategy):
             min_periods=self.params.slow,
         ).mean()
 
-        # базовый сигнал: 1 если fast > slow, иначе 0
-        target = (fast > slow).astype(int)
+        # базовый сигнал: сейчас 1 если fast > slow, иначе 0; 
+        # но в принципе может быть дробным (например, входим в BTC на 0.02 от капитала)
+        target = (fast > slow).astype(float)
 
         # лаг исполнения (на следующий бар)
         if self.params.shift_for_execution != 0:
             target = target.shift(self.params.shift_for_execution)
 
         # на начальных барах и после shift → flat
-        target = target.fillna(0).astype(int)
+        target = target.fillna(0.0).astype(float)
 
         return Signals(target_position=target)
